// dashboard/src/utils/hasura-api/generated/schemas/logicalModelSelectPermission.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { LogicalModelSelectPermissionFilter } from './logicalModelSelectPermissionFilter';

export interface LogicalModelSelectPermission {
  columns: '*' | string[];
  filter: LogicalModelSelectPermissionFilter;
}